requirements.txt
numpy==2.3.0
opencv_python==4.11.0.86
pycryptodome==3.23.0
pymodbus==3.9.2
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
